requirements.txt
python-telegram-bot==20.7
easyocr==1.7.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
